services/fits-metadata/requirements.txt
grpcio==1.76.0
protobuf==6.33.4
astropy==6.1.2
